requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
appdirs==1.4.4            # via black
arrow==0.15.6             # via -r requirements.in
astroid==2.4.1            # via pylint
attrs==19.3.0             # via -r requirements.in, black, pytest
black==19.10b0            # via -r requirements.in
bokeh==2.0.2              # via -r requirements.in
certifi==2020.4.5.1       # via requests
chardet==3.0.4            # via requests
click==7.1.2              # via -r requirements.in, black, pip-tools
coverage==5.1             # via pytest-cov
decorator==4.4.2          # via validators
docutils==0.16            # via m2r
flake8==3.8.2             # via -r requirements.in
idna==2.9                 # via requests
importlib-metadata==1.6.0  # via flake8, pluggy, pytest
isort==4.3.21             # via pylint
jinja2==2.11.2            # via bokeh
lazy-object-proxy==1.4.3  # via astroid
m2r==0.2.1                # via -r requirements.in
markupsafe==1.1.1         # via jinja2
mccabe==0.6.1             # via flake8, pylint
mistune==0.8.4            # via m2r
more-itertools==8.3.0     # via pytest
mypy-extensions==0.4.3    # via mypy
mypy==0.770               # via -r requirements.in
numpy==1.18.4             # via -r requirements.in, bokeh
packaging==20.4           # via bokeh, pytest
pathspec==0.8.0           # via black
pillow==7.1.2             # via bokeh
pip-tools==5.2.0          # via -r requirements.in
pluggy==0.13.1            # via pytest
py==1.8.1                 # via pytest
pycodestyle==2.6.0        # via flake8
pyflakes==2.2.0           # via flake8
pylint==2.5.2             # via -r requirements.in
pyparsing==2.4.7          # via packaging
pytest-cov==2.9.0         # via -r requirements.in
pytest==5.4.2             # via -r requirements.in, pytest-cov
python-dateutil==2.8.1    # via arrow, bokeh
pyyaml==5.3.1             # via -r requirements.in, bokeh
regex==2020.5.14          # via black
requests==2.23.0          # via -r requirements.in
six==1.15.0               # via astroid, packaging, pip-tools, python-dateutil, validators
toml==0.10.1              # via black, pylint
toolz==0.10.0             # via -r requirements.in
tornado==6.0.4            # via bokeh
typed-ast==1.4.1          # via astroid, black, mypy
typing-extensions==3.7.4.2  # via bokeh, mypy
urllib3==1.25.9           # via requests
validators==0.15.0        # via -r requirements.in
wcwidth==0.1.9            # via pytest
wrapt==1.12.1             # via astroid
zipp==3.1.0               # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# pip
